Added extra failing check to generation test
When a value is added with the max value for 24 bits, the next
generation bump will overflow generation ending up with a new 24 bit
generation value of 1.

This leaves the value with the max-24 generation with a generation that
satisfies none of the checks for Get. So elements of this generation
become unreadable.