Fix generation overflow (#17)
* Demonstrate Set/Get failure at 2^24 generations

There are only 24 bits available to carry the generation info in the
index map. When a bucket's generation is greater than 2^24 (16,777,216)
values added to the cache cannot be read from the cache at all.

* Read bucket generation as 24 bit value

If we always read a bucket's generation as a 24 bit value then the
generation value added to an element's index can continue to be compared
after generation overlows 24 bits.

* Added extra failing check to generation test

When a value is added with the max value for 24 bits, the next
generation bump will overflow generation ending up with a new 24 bit
generation value of 1.

This leaves the value with the max-24 generation with a generation that
satisfies none of the checks for Get. So elements of this generation
become unreadable.

* Checking for case where gen overflows in Get

If a value exists in the cache with the maximum value for 24 bits, the
next generation bump will give a generation value of 1.

We allow Gets of values with max-24 generation if the bucket's
generation is 1.