all: pass config as a parameter for init function
The rationale behind this is: in some case, the hasher
function may required to be overriden. XXhash is a very
fast hash function, but sometime there is no necessary
to re-compute the digest of key if the key itself is already
hashed.

For example the key itself is a 32byte hashed value, so
a simple Hasher function may just use the first 8 bytes
as the digest. Although it may increase the prosibility
of hash collision but it's faster.