Revert "introduce new stats item EvictedBytes (#93)"
This reverts commit 1412165ee6d53badc62cd4bab07b50c055cd0e10.

Reason for revert: the exposed EvictedBytes metric is misleading - it doesn't reflect the real
eviction rate of the data in the cache, since old data is overwritten by the new data
in the cyclic buffer formed from the cache chunks, on the addition of new items to the cache.
It is impossible to detect the eviction rate since the code doesn't know when it overwrites
the already existing data. The EvictedBytes metric is just increased by the total size of chunks
on every wraparound in the shard. This can be used as a rough estimation of the evicted bytes,
but it is updated too rarely on large caches.

It is possible to add WrittenBytes metric, which will show the number of bytes written to the cache.
This metric roughly equals to the EvictedBytes metric after the cache becomes full and old data
starts to be overwritten by the new data.

Both metrics have zero practical applicability, so it is better to drop the EvictedBytes metric
even if it has been already exposed in public, in order to reduce the complexity of the code and reduce
the confusion for the users who try using this metric. All the users of this metric must stop using it.

Updates https://github.com/VictoriaMetrics/fastcache/pull/93
Updates https://github.com/VictoriaMetrics/VictoriaMetrics/pull/9293